FastAPIServer/requirements.txt
fastapi[standard]
uvicorn
pdfplumber
databases
sqlalchemy
python-dotenv
asyncpg
psycopg2-binary